+++ components/templates/src/global_fns.rs
@@ -139,6 +139,7 @@ pub fn make_get_taxonomy(all_taxonomies: Vec<Taxonomy>) -> GlobalFn {
     for taxonomy in all_taxonomies {
         taxonomies.insert(taxonomy.kind.name.clone(), taxonomy);
     }
+
     Box::new(move |args| -> Result<Value> {
         let kind = required_arg!(
             String,
@@ -153,6 +154,42 @@ pub fn make_get_taxonomy(all_taxonomies: Vec<Taxonomy>) -> GlobalFn {
         };
 
         return Ok(to_value(container).unwrap());
+    })
+}
+
+pub fn make_get_taxonomy_url(all_taxonomies: Vec<Taxonomy>) -> GlobalFn {
+    let mut taxonomies = HashMap::new();
+    for taxonomy in all_taxonomies {
+        taxonomies.insert(taxonomy.kind.name.clone(), taxonomy);
+    }
+
+    Box::new(move |args| -> Result<Value> {
+        let kind = required_arg!(
+            String,
+            args.get("kind"),
+            "`get_taxonomy_url` requires a `kind` argument with a string value"
+        );
+        let name = required_arg!(
+            String,
+            args.get("name"),
+            "`get_taxonomy_url` requires a `name` argument with a string value"
+        );
+        let container = match taxonomies.get(&kind) {
+            Some(c) => c,
+            None => return Err(
+                format!("`get_taxonomy_url` received an unknown taxonomy as kind: {}", kind).into()
+            )
+        };
+
+        for item in &container.items {
+            if item.name == name {
+                return Ok(to_value(item.permalink.clone()).unwrap());
+            }
+        }
+
+        Err(
+            format!("`get_taxonomy_url`: couldn't find `{}` in `{}` taxonomy", name, kind).into()
+        )
     })
 }
 
@@ -206,7 +243,7 @@ pub fn make_resize_image(imageproc: Arc<Mutex<imageproc::Processor>>) -> GlobalF
 
 #[cfg(test)]
 mod tests {
-    use super::{make_get_url, make_get_taxonomy, make_trans};
+    use super::{make_get_url, make_get_taxonomy, make_get_taxonomy_url, make_trans};
 
     use std::collections::HashMap;
 
@@ -260,7 +297,7 @@ mod tests {
     fn can_get_taxonomy() {
         let taxo_config = TaxonomyConfig { name: "tags".to_string(), ..TaxonomyConfig::default() };
         let tag = TaxonomyItem::new(
-            "Prog amming",
+            "Progamming",
             "tags",
             &Config::default(),
             vec![],
@@ -278,6 +315,33 @@ mod tests {
         // and errors if it can't find it
         let mut args = HashMap::new();
         args.insert("kind".to_string(), to_value("something-else").unwrap());
+        assert!(static_fn(args).is_err());
+    }
+
+    #[test]
+    fn can_get_taxonomy_url() {
+        let taxo_config = TaxonomyConfig { name: "tags".to_string(), ..TaxonomyConfig::default() };
+        let tag = TaxonomyItem::new(
+            "Programming",
+            "tags",
+            &Config::default(),
+            vec![],
+        );
+        let tags = Taxonomy {
+            kind: taxo_config,
+            items: vec![tag],
+        };
+
+        let static_fn = make_get_taxonomy_url(vec![tags.clone()]);
+        // can find it correctly
+        let mut args = HashMap::new();
+        args.insert("kind".to_string(), to_value("tags").unwrap());
+        args.insert("name".to_string(), to_value("Programming").unwrap());
+        assert_eq!(static_fn(args).unwrap(), "http://a-website.com/tags/prog-amming/");
+        // and errors if it can't find it
+        let mut args = HashMap::new();
+        args.insert("kind".to_string(), to_value("tags").unwrap());
+        args.insert("name".to_string(), to_value("random").unwrap());
         assert!(static_fn(args).is_err());
     }